Feat: Update Makefile and docker-compose for authentication
Update the web server container name to "webserver-auth" in the Makefile and docker-compose.yaml file for identification of servers with authentication. Removed the 'push' trigger from the build.yml GitHub workflow file to prevent erroneous deployments on non-main branches. Changed exposed webserver port to 81 for better security. A new command 'make deps' is added to the makefiles to ensure dependencies are managed in the build process.