Feat: Removed `ClientRedirectUris.php` and `ClientScopes.php` classes and added the functionality in the `Client.php` class
Two domains `ClientRedirectUris` and `ClientScopes` were deleted. The functionality is brought into the `Client` class itself. This reduces the number of classes and simplifies the architecture. In addition, unused `val` variables in `ClientGrantType.php` and `ClientScopeType.php` are replaced with actual type hinting. Lastly, the `value` method in `ClientIdentifier.php` was given functionality.