Commit message:
Add network-auth and switch to postgres in Docker

Added a new network 'network-auth' to the Docker services auth and database-auth for isolating them to the new network. This allows for communication between services using their container name. Also switched the database from MySQL to PostgreSQL. This includes changes in Docker config for pgsql extension instead of mysql and an update in DATABASE_URL in .env file. Made changes in doctrine.yaml to explicitly state the server version as '15'.