Feat: Add Faker library to dev dependencies
Fakerphp/Faker library has been added to the dev dependencies for generating fake data for testing. This will help in better unit testing by providing different sets of data and scenarios. The library has been included in composer.json and also updated the composer.lock file to reflect the changes. The version chosen is compatible with the existing setup and does not interfere with other dependencies.